@@ -1,24 +1,18 @@
 const htmlStandards = require('reshape-standard')
 const cssStandards = require('spike-css-standards')
-const jsStandards = require('babel-preset-latest')
+const jsStandards = require('spike-js-standards')
 const pageId = require('spike-page-id')
+
 module.exports = {
   devtool: 'source-map',
   matchers: {
     html: '*(**/)*.sgr',
     css: '*(**/)*.sss'
   },
-  ignore: ['**/layout.sgr', '**/_*', '**/.*', '_cache/**', 'readme.md'],
-  reshape: (ctx) => {
-    return htmlStandards({
-      webpack: ctx,
-      locals: { pageId: pageId(ctx), foo: 'bar' }
-    })
-  },
-  postcss: (ctx) => {
-    const css = cssStandards({webpack: ctx})
-    css.plugins.push(require('lost')(), require('postcss-mixins')(), require('postcss-breakpoints')())
-    return css
-  },
-  babel: { presets: [jsStandards] }
+  ignore: ['**/layout.sgr', '**/_*', '**/.*', 'readme.md', 'yarn.lock'],
+  reshape: htmlStandards({
+    locals: (ctx) => { return { pageId: pageId(ctx), foo: 'bar' } }
+  }),
+  postcss: cssStandards(),
+  babel: jsStandards()
 }
